# executor: optimize UnionScanExec, try to use table/index access range to avoid execute buildAndSortAddedRows.