--- a/executor/builder.go
+++ b/executor/builder.go
@@ -804,6 +804,10 @@ func (b *executorBuilder) buildUnionScanFromReader(reader Executor, v *plannerco
 		us.dirty = GetDirtyDB(b.ctx).GetDirtyTable(physicalTableID)
 		us.conditions = v.Conditions
 		us.columns = x.columns
+		err = us.buildAddedRowsFromIndexReader(x)
+		if err != nil {
+			return nil, err
+		}
 		err = us.buildAndSortAddedRows(x.table)
 	case *IndexLookUpExecutor:
 		us.desc = x.desc