--- a/executor/union_scan.go
+++ b/executor/union_scan.go
@@ -16,21 +16,18 @@ package executor
 import (
 	"bytes"
 	"context"
-	"fmt"
-	"github.com/pingcap/tidb/distsql"
-	"github.com/pingcap/tidb/planner/core"
-	"github.com/pingcap/tidb/tablecodec"
-	"github.com/pingcap/tidb/util/codec"
 	"sort"
 	"time"
 
 	"github.com/opentracing/opentracing-go"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/parser/model"
+	"github.com/pingcap/tidb/distsql"
 	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/table"
 	"github.com/pingcap/tidb/table/tables"
+	"github.com/pingcap/tidb/tablecodec"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/chunk"
 )
@@ -216,7 +213,6 @@ func (us *UnionScanExec) getSnapshotRow(ctx context.Context) ([]types.Datum, err
 				continue
 			}
 			if _, ok := us.dirty.addedRows[snapshotHandle]; ok {
-				fmt.Printf("\n\nhandle: %v\n--------\n-n", snapshotHandle)
 				// If src handle appears in added rows, it means there is conflict and the transaction will fail to
 				// commit, but for simplicity, we don't handle it here.
 				continue
@@ -300,52 +296,61 @@ func (us *UnionScanExec) rowWithColsInTxn(t table.Table, h int64, cols []*table.
 	return v, nil
 }
 
-func (us *UnionScanExec) buildAddedRowsFromIndexReader(e *IndexReaderExecutor) error {
-	condition := us.conditions
-	if len(e.plans) > 1 {
-		if sel, ok := e.plans[1].(*core.PhysicalSelection); ok {
-			condition = sel.Conditions
+func (us *UnionScanExec) checkBypassAddedRowsByTableReader(e *TableReaderExecutor) bool {
+	if len(e.ranges) == 0 {
+		return false
+	}
+	physicalTableID := getPhysicalTableID(e.table)
+	kvRanges := distsql.TableRangesToKVRanges(physicalTableID, e.ranges, nil)
+	txn, err := us.ctx.Txn(true)
+	if err != nil {
+		return false
+	}
+
+	prefix := tablecodec.GenTableRecordPrefix(physicalTableID)
+	for _, rg := range kvRanges {
+		iter, err := txn.GetMemBuffer().Iter(rg.StartKey, rg.EndKey)
+		if err != nil {
+			return false
+		}
+		for iter.Valid() {
+			if !bytes.Contains(iter.Key(), prefix) {
+				continue
+			}
+			return false
 		}
 	}
-	_ = condition
+	return true
+}
+
+func (us *UnionScanExec) checkBypassAddedRowsByIndexReader(e *IndexReaderExecutor) bool {
+	if len(e.ranges) == 0 {
+		return false
+	}
 	kvRanges, err := distsql.IndexRangesToKVRanges(e.ctx.GetSessionVars().StmtCtx, e.physicalTableID, e.index.ID, e.ranges, e.feedback)
 	if err != nil {
-		return err
+		return false
 	}
 
 	txn, err := us.ctx.Txn(true)
 	if err != nil {
-		return err
+		return false
 	}
 
 	prefix := tablecodec.EncodeTableIndexPrefix(e.physicalTableID, e.index.ID)
 	for _, rg := range kvRanges {
 		iter, err := txn.GetMemBuffer().Iter(rg.StartKey, rg.EndKey)
 		if err != nil {
-			return err
+			return false
 		}
 		for iter.Valid() {
-			key := iter.Key()
-			value := iter.Value()
-			if !bytes.Contains(key, prefix) {
+			if !bytes.Contains(iter.Key(), prefix) {
 				continue
 			}
-			key = key[len(prefix):]
-			fmt.Printf("\n\nkey: %v, value: %v\n\n------------------\n", key, value)
-
-			ds, err := codec.Decode(key, len(e.index.Columns)+1)
-			if err != nil {
-				return err
-			}
-			fmt.Printf("\n\nds %#v\n\n------------------\n", ds)
-
-			err = iter.Next()
-			if err != nil {
-				return err
-			}
+			return false
 		}
 	}
-	return nil
+	return true
 }
 
 func (us *UnionScanExec) buildAndSortAddedRows(t table.Table) error {