<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
In the case of big transaction, `buildAndSortAddedRows` function will spends too much time to decode and sort data.  

This PR try to avoid call `buildAndSortAddedRows` function in in some scenarios. here is a example:

```sql
 create table t (a int key, b int, index idx(b));
 insert into t values (1,1)
 begin
 insert into t values (2,2),(3,3),(4,4),(5,5)
 update t set b=b+1 where b=1;
 commit
```

In the upper transaction, execute `update t set b=b+1 where b=1;` will use `UnionScanExec` with `IndexReaderExecutor` and the index range is `range:[1,1]`.

Here is a trick, use the index `range:[1,1]` to check `MemBuffer` first, and if there is no row in `MemBuffer`, then we can avoid call `buildAndSortAddedRows`.

## benchmark

Below is a simple benchmark, Only 1 TiDB server, no pd/tikv.

**table schema**
```sql
create table t (a int key, b int,c varchar(100),d varchar(100), e decimal(60,10), f timestamp, g double, index idx(b));
```

***benchmark test***

**test insert 100 rows in transaction.**
```sql
begin
# insert 100 rows data.
insert into t2 values (0,0,'abcdefg', 'hijklmn', 123.456, '2019-05-29 20:16:59', 0.6046602879796196) ... # 100 rows data.
select * from t where a>5000; # do a benchmark of this select. the select result is null. 
rollback
```

In Master, the select avg spend time is **0.41ms**

In This PR, the select avg spend time is **0.22ms**

**test insert 1000 rows in transaction.**

In Master, the select avg spend time is **2ms**

In This PR, the select avg spend time is **0.22ms**

Here is the cpu-profile:
[cpu.profile-PR.txt](https://github.com/pingcap/tidb/files/3232594/cpu.profile-PR.txt)
[cpu.profile-master.txt](https://github.com/pingcap/tidb/files/3232600/cpu.profile-master.txt)




## Beter Idea

How about use `TableReaderExecutor` and `IndexReaderExecutor` to read `MemBuffer` and remove `buildAndSortAddedRows` function?



### What is changed and how it works?
* use table/index access range to avoid execute buildAndSortAddedRows.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->


Code changes

 - Has exported function/method change

Side effects


Related changes