@crazycs520 2019-05-29
Maybe a better idea is implement a `TableReaderExecutor`, `IndexReaderExecutor` and `IndexLookUpExecutor` to read `MemBuffer` and remove `buildAndSortAddedRows` function.